Move `TypeTree` classes inside `Trees`.
And remove `TypeTrees`.
