+++ tasty-query/shared/src/main/scala/tastyquery/reader/tasties/TreeUnpickler.scala
@@ -19,7 +19,6 @@ import tastyquery.Spans.*
 import tastyquery.Symbols.*
 import tastyquery.Trees.*
 import tastyquery.Types.*
-import tastyquery.TypeTrees.*
 
 import TastyUnpickler.NameTable
